pyproject.toml
[tool.poetry]
name = "space-station"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "space_station"}]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.16"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"
djangorestframework-simplejwt = "^4.7.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

